Add JavaDoc to core and task package files
