+++ src/main/java/myapp/core/BingBongBot.java
@@ -3,11 +3,21 @@ package myapp.core;
 import myapp.command.Command;
 import myapp.task.TaskList;
 
+/**
+ * Represents the BingBongBot that handles user interactions, manages tasks, and
+ * processes commands.
+ */
 public class BingBongBot {
     private final BingBongUI ui;
     private final Storage storage;
     private TaskList tasks;
 
+    /**
+     * Constructs a BingBongBot with the specified UI and storage.
+     *
+     * @param ui the UI to interact with the user.
+     * @param storage the storage system to load and save tasks.
+     */
     public BingBongBot(BingBongUI ui, Storage storage) {
         this.ui = ui;
         this.storage = storage;
@@ -18,6 +28,10 @@ public class BingBongBot {
         }
     }
 
+    /**
+     * Starts the BingBongBot, allowing it to run and process user commands until
+     * the user issues an exit command.
+     */
     public void run() {
         ui.showGreeting();
         boolean isExit = false;
@@ -33,7 +47,7 @@ public class BingBongBot {
             } catch (NumberFormatException e) {
                 ui.showResponse("Invalid task number. Please enter a valid number.");
             } catch (IndexOutOfBoundsException e) {
-                ui.showResponse("myapp.task.Task number is out of range. Please enter a valid task number.");
+                ui.showResponse("Task number is out of range. Please enter a valid task number.");
             } catch (Exception e) {
                 ui.showResponse("An unexpected error occurred: " + e.getMessage());
             }
