+++ src/main/java/myapp/core/Storage.java
@@ -7,13 +7,30 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
 
+/**
+ * The {@code Storage} class handles the loading and saving of tasks to and from a file.
+ * It provides methods to read tasks from a file into a {@link TaskList} and save tasks
+ * from a {@link TaskList} to a file.
+ */
 public class Storage {
     private final String filePath;
 
+    /**
+     * Constructs a {@code Storage} object with the specified file path.
+     *
+     * @param filePath the file path where tasks will be loaded from and saved to.
+     */
     public Storage(String filePath) {
         this.filePath = filePath;
     }
 
+    /**
+     * Loads tasks from the specified file.
+     *
+     * @return a list of {@link Task} objects loaded from the file.
+     * @throws BingBongException if an error occurs while reading from the file
+     * or if the file contains unknown task types.
+     */
     public List<Task> load() throws BingBongException {
         List<Task> tasks = new ArrayList<>();
 
@@ -28,7 +45,7 @@ public class Storage {
                 Task task;
                 switch (type) {
                 case "T":
-                    task = new Todo(description);
+                    task = new ToDo(description);
                     break;
                 case "D":
                     String by = taskData[3];
@@ -53,6 +70,12 @@ public class Storage {
         return tasks;
     }
 
+    /**
+     * Saves the tasks in the given {@link TaskList} to the specified file.
+     *
+     * @param tasks the {@link TaskList} containing the tasks to be saved.
+     * @throws BingBongException if an error occurs while writing to the file or if the directory cannot be created.
+     */
     public void save(TaskList tasks) throws BingBongException {
         File file = new File(filePath);
