--- src/main/java/myapp/task/Todo.java
@@ -1,18 +1,0 @@
-package myapp.task;
-
-public class Todo extends Task{
-
-    public Todo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-
-    @Override
-    public String toFileFormat() {
-        return "T | " + (isDone ? "1" : "0") + " | " + description;
-    }
-}
