+++ src/test/java/myapp/task/ToDoTest.java
@@ -4,18 +4,18 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class TodoTest {
+public class ToDoTest {
 
     @Test
     public void testStringConversion(){
         assertEquals("[T][ ] go buy groceries",
-                new Todo("go buy groceries").toString());
+                new ToDo("go buy groceries").toString());
     }
 
     @Test
     public void testFileDataConversion(){
         assertEquals("T | 0 | go buy groceries",
-                new Todo("go buy groceries").toFileFormat());
+                new ToDo("go buy groceries").toFileFormat());
     }
